Fix bash-completion for git 1.7.2.
The completion for git 1.7.2 parses the output of `git help --all`
to generate a list of available commands. The output from `hub help`
does not conform to the same format or respect the `--all` flag.

To fix this issue, we only display hub's extended help when the
`--all` (or `-a`) flag is not present, falling back to git.